remove price from consultations

+++ db/migrate/20220712063405_remove_price_from_consultations.rb
@@ -1,5 +1,5 @@
 class RemovePriceFromConsultations < ActiveRecord::Migration[6.1]
   def change
-    remove_column :consultations, :price, :integer
+    remove_column :consultations, :price, :float
   end
 end
